bench: Simplify calls to legacy benchmarks
